requirements.txt
gunicorn
Flask
flask_sso
sqlalchemy
pycrypto
requests
psycopg2
pytz